% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{data_generation}
\alias{data_generation}
\title{=========@data_generation
=========@ Model Y.ij =X1.ij' beta1 +X2.ij' beta2 + Z1.i' alpha.i + error}
\usage{
data_generation(n_sites, n_fam_ps, n_ind_pf, n_vox, beta)
}
\description{
=========@data_generation
=========@ Model Y.ij =X1.ij' beta1 +X2.ij' beta2 + Z1.i' alpha.i + error
}
