% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permuteWparallel.R
\name{permute_and_fit}
\alias{permute_and_fit}
\title{=============== @permutesd_score_stats-with-parallelization}
\usage{
permute_and_fit(data, dependentVar, voxels, groupVar)
}
\description{
=============== @permutesd_score_stats-with-parallelization
}
